final working version

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: asplavni <asplavni@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 17:59:33 by asplavni          #+#    #+#             */
-/*   Updated: 2024/01/12 22:32:24 by asplavni         ###   ########.fr       */
+/*   Updated: 2024/01/13 11:10:32 by asplavni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,8 +24,8 @@ int		ft_putchar(char c);
 int		ft_putstr(char *s);
 size_t	ft_strlen(const char *s);
 
-int		ft_putdigit(long n, int base);
+int		ft_putdigit(long long n, int base);
 int		ft_hexa(long n, int base, char UP_or_low);
-int		ft_pointer(long n, int base);
+int		ft_pointer(unsigned long n, int base);
 
 #endif
